+++ ledger/models/characteraudit.py
@@ -16,7 +16,6 @@ from allianceauth.services.hooks import get_extension_logger
 # Alliance Auth (External Libs)
 from eve_sde.models import ItemType
 from eve_sde.models.map import SolarSystem
-from eveuniverse.models import EveMarketPrice
 
 # AA Ledger
 from ledger import __title__
@@ -32,6 +31,7 @@ from ledger.managers.character_audit_manager import (
 from ledger.managers.character_journal_manager import CharWalletManager
 from ledger.managers.character_mining_manager import CharacterMiningLedgerEntryManager
 from ledger.models.general import (
+    EveMarketPrice,
     UpdateSectionResult,
     UpdateStatusBaseModel,
     WalletJournalEntry,
@@ -142,17 +142,17 @@ class CharacterOwner(models.Model):
         )
 
     @property
-    def mining_ledger(self):
+    def mining_ledger(self) -> models.QuerySet["CharacterMiningLedger"]:
         """Get the mining ledger for this character."""
         return self.ledger_character_mining
 
     @property
-    def wallet_journal(self):
+    def wallet_journal(self) -> models.QuerySet["CharacterWalletJournalEntry"]:
         """Get the wallet journal for this character."""
         return self.ledger_character_journal
 
     @property
-    def update_status(self):
+    def update_status(self) -> models.QuerySet["CharacterUpdateStatus"]:
         return self.ledger_update_status
 
     def get_portrait(self, size: int = 64, as_html: bool = False) -> str:
@@ -270,7 +270,7 @@ class CharacterMiningLedger(models.Model):
         return f"{mining_record.date.strftime('%Y%m%d')}-{mining_record.type_id}-{character_id}-{mining_record.solar_system_id}"
 
     @staticmethod
-    def update_evemarket_price():  # Dont want to make a task only for this
+    def update_evemarket_price():
         """Update Prices for the EveMarketPrice."""
         updated = 0
         cached_update = cache.get(f"{LEDGER_CACHE_KEY}-eve-market-price", False)